Moved encounter generation to a seperate utils file
